+++ login_control.js
@@ -1,12 +1,16 @@
 /**
- * Login Control plugin script
+ * Roundcube Plugin Login Control
+ * Plugin to add whitelist and blacklist for login.
  *
- * @version 1.2
+ * @version 1.3
  * @author Alexander Pushkin
- * @url https://github.com/san4op/roundcube_login_control
- * @licence GNU GPLv3
+ * @copyright Copyright (c) 2017, Alexander Pushkin
+ * @link https://github.com/san4op/roundcube_login_control
+ * @license GNU General Public License, version 3
  */
 
-rcmail.addEventListener('plugin.access_restricted', function(response){
-	rcmail.command('logout');
-});
+if (window.rcmail) {
+	rcmail.addEventListener('plugin.access_restricted', function(response) {
+		rcmail.command('switch-task', 'logout');
+	});
+}