+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+RELEASE 1.3
+-----------
+* Configuration moved to local config.inc.php.
+* Added global list: use * (asterisk) as username.
+* Small refactoring.
+* Update translations.
+
 RELEASE 1.2
 -----------
 * Rename plugin